Neon 2 and 3 merged together
Neon2 is loaded and users can click a button to enter edit mode and edit the mei file.
