+++ package.json
@@ -38,7 +38,8 @@
         "express": "^4.16.3",
         "multer": "^1.3.0",
         "nodemon": "^1.17.4",
-        "pug": "^2.0.3"
+        "pug": "^2.0.3",
+        "webpack": "^4.12.0"
     },
     "scripts": {
         "start": "nodemon server.js",
@@ -50,7 +51,6 @@
         "d3": "^5.5.0",
         "jquery": "^3.3.1",
         "style-loader": "^0.21.0",
-        "webpack": "^4.12.0",
         "webpack-cli": "^3.0.8"
     }
 }
