--- public/src/EditControls.js
@@ -1,1 +1,0 @@
-function EditControls () {}
