+++ src/EditControls.js
@@ -1,0 +1,3 @@
+function EditControls () {}
+
+export { EditControls };
