+++ src/Navbar.js
@@ -1,4 +1,4 @@
-function Navbar(filename) {  
+function Navbar (filename) {  
     // setup navbar listeners 
     $("#revert").on("click", function(){
         if (confirm("Reverting will cause all changes to be lost. Press OK to continue.")) {
@@ -10,9 +10,11 @@ function Navbar(filename) {
     });
 
     //mei download link
-    $("#getmei").attr("href", '/../uploads/mei/' + filename);
+    $("#getmei").attr("href", filename);
 
     //png download setup
     var pngFile = filename.split('.', 2)[0] + ".png";
-    $("#getpng").attr("href", '/../uploads/png/' + pngFile);
+    $("#getpng").attr("href", pngFile);
 }
+
+export { Navbar };
