+++ src/editor.js
@@ -2,6 +2,7 @@ import { Neon } from './Neon.js';
 import 'bulma/css/bulma.css';
 import './bulma-slider.min.css';
 import './bulma-slider.min.js';
+import '../public/css/style.css'
 
 var neon = new Neon({
     meifile: meiFile,
