+++ views/editor.pug
@@ -13,7 +13,7 @@ html
       .column#svg_output.is-two-thirds.box(style={overflow: "hidden"})
       .column.is-one-third
         .panel
-          p.panel-heading Controls
+          p.panel-heading Display
           a.panel-block.has-text-centered
             button.button#reset-zoom Zoom
             input.slider.is-fullwidth#zoomSlider(step="5", min="25", max="200", value="100", type="range")
@@ -29,6 +29,7 @@ html
           a.panel-block
             label.checkbox Display Text: 
               input.checkbox#displayText(type="checkbox")
+          #edit_controls
         #neume_info
       script(type='text/javascript').
         var meiFile = "#{meifile}";
